ml_project/requirements.txt
async-generator==1.10
async-timeout==3.0.1
backports.functools-lru-cache==1.6.1
blinker==1.4
brotlipy==0.7.0
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorlog==4.7.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
google-auth==1.30.0
google-auth-oauthlib==0.4.1
idna==2.10
joblib==1.0.1
marshmallow==3.11.1
marshmallow-dataclass==8.4.1
mypy-extensions==0.4.3
numpy==1.20.2
oauthlib==3.0.1
olefile==0.46
pandas==1.2.4
pandocfilters==1.4.2
protobuf==3.15.5
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
PyYAML==5.4.1
pyzmq==19.0.2
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
Send2Trash==1.5.0
six==1.16.0
sklearn==0.0
testpath==0.4.4
threadpoolctl==2.1.0
typing-extensions==3.10.0.0
typing-inspect==0.6.0
urllib3==1.26.4
webencodings==0.5.1
Werkzeug==1.0.1
